added location buttons, passed lat long as props. functional test
