+++ renewable-energy-dashboard/src/index.css
@@ -35,7 +35,22 @@ code {
     height: 100%;
     width: 50%;
   }
- 
+.section-top {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 20vh;
+  width: 100%;
+  margin: 0 auto;
+}
+.location-buttons-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 10vh;
+width:80%
+
+}
 .button-container {
   display: flex;
   justify-content: center;
@@ -62,6 +77,7 @@ code {
 }
 .location-button {
   padding: 12px 24px;
+  margin: 1vh;
   background: linear-gradient(to right, #87CEEB, #00BFFF);
   border: none;
   color: white;
